Move Sidebar and VersionSelector under DocsLayout
