+++ src/components/layout/DocsLayout/index.ts
@@ -1,0 +1,3 @@
+import DocsLayout from './DocsLayout';
+
+export default DocsLayout;
