+++ src/components/screens/DocsScreen/VersionCTA.stories.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { styled } from '@storybook/theming';
 import { VersionCTA } from './VersionCTA';
-import { pageContext } from '../../layout/DocsLayout.stories';
+import { pageContext } from '../../layout/DocsLayout/DocsLayout.stories';
 
 const { versions } = pageContext;
